WIP: split image proxy contracts for LOC policy (checkpoint)
