+++ Services/ImageProxyService.cs
@@ -5,76 +5,6 @@ using Wayfarer.Util;
 namespace Wayfarer.Services;
 
 /// <summary>
-/// Contract for fetching, optimizing, and caching external images.
-/// Used by the cache warm-up job and any component that needs to pre-populate
-/// the proxied image cache without going through the HTTP controller pipeline.
-/// </summary>
-public interface IImageProxyService
-{
-    /// <summary>
-    /// Gets a proxied image from cache or, when allowed, from the origin.
-    /// Local cache hits schedule stale refresh as needed before origin rate limits apply.
-    /// </summary>
-    Task<ImageProxyResult> GetOrFetchAsync(ImageProxyRequest request, bool allowOriginFetch, CancellationToken ct = default);
-
-    /// <summary>
-    /// Forces origin refresh for a cache key through the shared per-key coordinator and process budget.
-    /// Used by bounded stale refresh and warm-up paths.
-    /// </summary>
-    Task<ImageProxyResult> RefreshAsync(ImageProxyRequest request, CancellationToken ct = default);
-
-    /// <summary>
-    /// Fetches an external image, optimizes it via ImageSharp, and stores it in the
-    /// proxied image cache. Returns true if the image was fetched and cached, false
-    /// if already cached, disallowed, or failed.
-    /// </summary>
-    /// <param name="imageUrl">The external image URL to fetch and cache.</param>
-    /// <param name="ct">Cancellation token.</param>
-    /// <returns>True if newly cached, false otherwise.</returns>
-    Task<bool> FetchAndCacheAsync(string imageUrl, CancellationToken ct = default);
-}
-
-/// <summary>
-/// Immutable request parameters that determine a proxied image cache key and output bytes.
-/// </summary>
-public sealed record ImageProxyRequest(
-    string Url,
-    int? MaxWidth = null,
-    int? MaxHeight = null,
-    int? Quality = null,
-    bool Optimize = true);
-
-/// <summary>
-/// Result status for proxied image service work.
-/// </summary>
-public enum ImageProxyResultStatus
-{
-    FreshHit,
-    StaleHit,
-    Fetched,
-    OriginRequired,
-    BadRequest,
-    NotFound,
-    TooLarge,
-    Failed
-}
-
-/// <summary>
-/// Result returned by the proxied image pipeline.
-/// </summary>
-public sealed record ImageProxyResult(
-    ImageProxyResultStatus Status,
-    string CacheKey,
-    byte[]? Bytes,
-    string? ContentType)
-{
-    /// <summary>
-    /// Gets whether bytes are available to serve to the HTTP caller.
-    /// </summary>
-    public bool HasBytes => Bytes is { Length: > 0 } && !string.IsNullOrWhiteSpace(ContentType);
-}
-
-/// <summary>
 /// Fetches external images, optimizes them via ImageSharp, and stores them in the
 /// proxied image disk cache. Delegates to <see cref="ImageProxyHelper"/> for
 /// shared SSRF checks, cache key computation, and image optimization.
