+++ Services/ProxiedImageCacheService.cs
@@ -5,77 +5,6 @@ using Wayfarer.Parsers;
 namespace Wayfarer.Services;
 
 /// <summary>
-/// Defines the contract for caching proxied and optimized external images on disk.
-/// Provides disk-based caching with DB-tracked LRU eviction, following the same
-/// pattern as <see cref="TileCacheService"/>.
-/// Read operations are lock-free for performance; writes and eviction use a global lock.
-/// </summary>
-public interface IProxiedImageCacheService
-{
-    /// <summary>
-    /// Returns an explicit cache result for the requested key.
-    /// Fresh and stale hits include local bytes; misses and disk failures do not.
-    /// </summary>
-    Task<ProxiedImageCacheResult> GetAsync(string cacheKey);
-
-    /// <summary>
-    /// Stores processed image bytes under the given cache key.
-    /// Existing entries are atomically replaced so readers see complete old or new bytes.
-    /// Triggers LRU eviction if the cache size limit would be exceeded.
-    /// </summary>
-    Task SetAsync(string cacheKey, byte[] bytes, string contentType);
-
-    /// <summary>
-    /// Ensures the cache directory exists and initializes cache size tracking from the database.
-    /// Called once at application startup.
-    /// </summary>
-    void Initialize();
-
-    /// <summary>
-    /// Returns the total cached image size in megabytes (from DB metadata).
-    /// </summary>
-    Task<double> GetCacheSizeInMbAsync();
-
-    /// <summary>
-    /// Returns the total number of cached images (from DB metadata).
-    /// </summary>
-    Task<int> GetCachedImageCountAsync();
-}
-
-/// <summary>
-/// Describes the cache state for a proxied image lookup.
-/// </summary>
-public enum ProxiedImageCacheStatus
-{
-    /// <summary>No metadata exists for the cache key.</summary>
-    Miss,
-
-    /// <summary>Metadata and a fresh local file exist.</summary>
-    FreshHit,
-
-    /// <summary>Metadata and a stale-but-servable local file exist.</summary>
-    StaleHit,
-
-    /// <summary>Metadata exists, but the local file is missing or unreadable.</summary>
-    DiskMissingOrError
-}
-
-/// <summary>
-/// Explicit result returned by proxied image cache reads.
-/// </summary>
-public sealed record ProxiedImageCacheResult(
-    ProxiedImageCacheStatus Status,
-    byte[]? Bytes,
-    string? ContentType,
-    string? FilePath)
-{
-    /// <summary>
-    /// Gets whether this result contains bytes that can be returned to the browser.
-    /// </summary>
-    public bool HasBytes => Bytes is { Length: > 0 };
-}
-
-/// <summary>
 /// Disk-based image proxy cache with DB-tracked LRU eviction.
 /// Caches optimized images from the ProxyImage endpoint to avoid repeated
 /// downloads and ImageSharp processing on every request.
